backend/requirements.txt
fastapi
fastapi[standard]
uvicorn
uvicorn[standard]
sqlalchemy
python-dotenv
pydantic
pytz